refactor(tenant-api): extract special-file writers into gitops/writer_special.go
Continue the writer.go split (Cycle 4 extracted gitcmd.go): move the non-tenant
entity write paths — WriteGroupsFile, WriteViewsFile, WriteFederationPolicyFile,
WriteFederationSubsetFile, and their shared writeSpecialFile helper — into a new
same-package file writer_special.go. writer.go drops 842 -> 757 lines; the
tenant write path and entity write paths now read separately.

Pure intra-package file move: same package gitops, no caller/test/signature
changes, no import changes to writer.go. Adversarial review confirms all 5
symbols conserved (declared once), bodies byte-identical, and — the key risk —
the admission-control ordering (MkdirAll -> acquireWrite -> w.mu.Lock) intact.
go build / vet / test green.

Self-Review-Pass-2: adversarial review: faithful move; 5 symbols conserved; ordering intact
